+++ qupath-extension-openslide/src/main/java/qupath/lib/images/servers/openslide/OpenslideImageServer.java
@@ -28,6 +28,7 @@ import java.awt.Graphics2D;
 import java.awt.image.BufferedImage;
 import java.awt.image.DataBufferInt;
 import java.io.IOException;
+import java.lang.ref.Cleaner;
 import java.net.URI;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -63,6 +64,30 @@ import qupath.lib.images.servers.openslide.jna.OpenSlideLoader;
 public class OpenslideImageServer extends AbstractTileableImageServer {
 	
 	private static final Logger logger = LoggerFactory.getLogger(OpenslideImageServer.class);
+
+	/**
+	 * Use a Cleaner to ensure that the OpenSlide object is closed when the server is no longer needed.
+	 * This is necessary because OpenSlide uses native resources that need to be released,
+	 * and finalize() is both a bad idea and deprecated for removal.
+	 */
+	private static final class OpenSlideState implements Runnable {
+
+		private final OpenSlide osr;
+
+		private OpenSlideState(OpenSlide osr) {
+			this.osr = osr;
+		}
+
+		@Override
+		public void run() {
+			logger.debug("Closing OpenSlide instance");
+			osr.close();
+		}
+	}
+
+	private static final Cleaner cleaner = Cleaner.create();
+	private final OpenSlideState state;
+	private final Cleaner.Cleanable cleanable;
 
 	private static boolean useBoundingBoxes = true;
 
@@ -121,6 +146,8 @@ public class OpenslideImageServer extends AbstractTileableImageServer {
 			osr = OpenSlideLoader.openImage(uri.toString());
 			name = null;
 		}
+		state = new OpenSlideState(osr);
+		cleanable = cleaner.register(this, state);
 
 		// Parse the parameters
 		int width = (int)osr.getLevel0Width();
@@ -261,9 +288,7 @@ public class OpenslideImageServer extends AbstractTileableImageServer {
 	
 	@Override
 	public void close() {
-		if (osr != null) {
-			osr.close();
-		}
+		cleanable.clean();
 	}
 
 	/**